refactor: XMLPreviewView.swift
